requirements/api.txt
fastapi==0.116.2
uvicorn==0.35.0
starlette==0.48.0
pydantic==2.11.9
pydantic_core==2.33.2

scikit-learn==1.7.2
numpy==2.3.3
pandas==2.3.2
joblib==1.5.2
scipy==1.16.2

pyyaml==6.0.3
python-dotenv==1.2.1

typing_extensions==4.15.0
anyio==4.10.0
h11==0.16.0
sniffio==1.3.1